Changed universal data types to Arduino specific ones

+++ config/config.cpp
@@ -4,7 +4,7 @@
  * Created Date: 14.11.2022 19:12:41
  * Author: 3urobeat
  * 
- * Last Modified: 15.11.2022 14:31:57
+ * Last Modified: 16.11.2022 19:03:00
  * Modified By: 3urobeat
  * 
  * Copyright (c) 2022 3urobeat <https://github.com/HerrEurobeat>
@@ -18,7 +18,7 @@
 // Define config variables
 namespace Config
 {
-    const int maxcol = 20; // Width of the display
+    const uint8_t maxcol = 20; // Width of the display
 
     const char *wifiSSID[2] = { "", "" }; // You can provide multiple networks if you wish
     const char *wifiPW[2]   = { "", "" };
@@ -33,10 +33,10 @@ namespace Config
     const char timeformat[9]       = "hh:mm:ss";
     const char miniClockFormat[6]  = "hh:mm";
 
-    const char *pageOrder[3]       = { "clock", "weather", "news" };
-    const int   showuntil[3]       = { 5000, 5000, 30000 }; // How long each page should be shown in ms
-    const bool  alwaysShowTime     = true; // Always show the time in the upper right corner
-    const int   clockWeekdaySwitch = 2500; // After how much ms the clock page should switch between date and weekday. Set to 0 to disable
+    const char    *pageOrder[3]       = { "clock", "weather", "news" };
+    const uint32_t showuntil[3]       = { 5000, 5000, 30000 }; // How long each page should be shown in ms
+    const bool     alwaysShowTime     = true; // Always show the time in the upper right corner
+    const uint16_t clockWeekdaySwitch = 2500; // After how much ms the clock page should switch between date and weekday. Set to 0 to disable
 
     const bool DEBUG = false;
 }
