+++ src/helpers/initWifi.cpp
@@ -4,7 +4,7 @@
  * Created Date: 30.08.2021 15:42:00
  * Author: 3urobeat
  * 
- * Last Modified: 15.11.2022 13:17:51
+ * Last Modified: 16.11.2022 19:15:00
  * Modified By: 3urobeat
  * 
  * Copyright (c) 2021 3urobeat <https://github.com/HerrEurobeat>
@@ -21,7 +21,7 @@
 void(* resetFunc) (void) = 0; // create a standard reset function
 
 
-ESP8266WiFiClass initWifi(size_t ssidamount, int row) 
+ESP8266WiFiClass initWifi(size_t ssidamount, uint8_t row) 
 {
     lcd.clearLine(row); //clear line just to make sure no old characters are left
 
@@ -30,14 +30,14 @@ ESP8266WiFiClass initWifi(size_t ssidamount, int row)
     delay(500);
 
     //Search for wifi networks in range
-    int  n     = WiFi.scanNetworks();
+    uint8_t n = WiFi.scanNetworks();
     bool found = false;
 
     char thisSSID[64]; //wifi SSIDs should have a max length of 32 characters
 
     //Try to connect by iterating for every found network over all networks in array (I tried doing this with indexOf() to avoid using a nested for loop but it had different results depending on the order in wifiSSID which was weird)
-    for (int i = 0; i <= n; i++) {
-        for (unsigned int j = 0; j < ssidamount; j++) {
+    for (uint8_t i = 0; i <= n; i++) {
+        for (uint8_t j = 0; j < ssidamount; j++) {
             WiFi.SSID(i).toCharArray(thisSSID, 64, 0);
 
             if (strcmp(thisSSID, Config::wifiSSID[j]) == 0) {
@@ -59,8 +59,8 @@ ESP8266WiFiClass initWifi(size_t ssidamount, int row)
         }
     }
 
-    //Block futher execution until connection is established and show message with waiting animation
-    int dots = 0;
+    // Block futher execution until connection is established and show message with waiting animation
+    uint8_t dots = 0;
 
     while (WiFi.status() != WL_CONNECTED)
     {
@@ -70,7 +70,7 @@ ESP8266WiFiClass initWifi(size_t ssidamount, int row)
         lcd.setCursor(13, row);
         lcd.print("    "); //Clear all dots
 
-        for (int i = 0; i < dots; i++) {
+        for (uint8_t i = 0; i < dots; i++) {
             lcd.setCursor(13 + i, row); 
             lcd.print("."); //add dots amount of dots behind "Connecting"
         }
