+++ src/loop.cpp
@@ -4,7 +4,7 @@
  * Created Date: 30.10.2022 19:01:32
  * Author: 3urobeat
  * 
- * Last Modified: 15.11.2022 17:55:53
+ * Last Modified: 16.11.2022 19:05:49
  * Modified By: 3urobeat
  * 
  * Copyright (c) 2022 3urobeat <https://github.com/HerrEurobeat>
@@ -19,8 +19,8 @@
 #include "pages/pages.h"
 
 
-int           currentPage = -1;      // Current page index in pageOrder array (init with  -1 to not skip page 0 on startup)
-int           oldPage;               // Save previous page to determine if we need to call lcd.clear() (useful if user only set one page in pageOrder to avoid blinking every showuntil seconds)
+int8_t        currentPage = -1;      // Current page index in pageOrder array (init with  -1 to not skip page 0 on startup)
+int8_t        oldPage;               // Save previous page to determine if we need to call lcd.clear() (useful if user only set one page in pageOrder to avoid blinking every showuntil seconds)
 unsigned long pageUpdate;            // Save timestamp when page was updated in order to keep track of showuntil without blocking thread with delay()
 bool          hideMiniClock = false; // Will be set to true when clockpage is active
